+++ packages/bluesky/lib/src/bluesky.dart
@@ -88,6 +88,9 @@ sealed class Bluesky {
         ),
       );
 
+  /// Returns the merged headers with global headers and auth header.
+  Map<String, String> get headers;
+
   /// Returns the current session.
   ///
   /// Set only if an instance of this object was created in
@@ -212,6 +215,9 @@ final class _Bluesky implements Bluesky {
         label = ctx.atproto.label,
         temp = ctx.atproto.temp,
         _ctx = ctx;
+
+  @override
+  Map<String, String> get headers => _ctx.headers;
 
   @override
   core.Session? get session => _ctx.session;